% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_and_summarise.R
\name{align_and_summarise}
\alias{align_and_summarise}
\title{Aligns sequence data, plots haplotype networks and summarises the data}
\usage{
align_and_summarise(alignment_files, max_haps_found_together, minbp)
}
\arguments{
\item{alignment_files}{list of .fas/.fasta files in directory that need
alignment, often signaled by the pattern "FOR_ALIGNMENT" in file name}

\item{max_haps_found_together}{input value for \code{hapfreq_from_paper} function.
Threshold value for number of times a haplotype can be found repeated in the
uploads from one study before paper is assumed to have submitted every sample
rather than unique haplotypes.}

\item{minbp}{minimum length (base pairs) for a sequence to be retained in the
data set}
}
\description{
\code{align_and_summarise}
}
\details{
Function aligns sequence data, plots two types of diagnostic plots (haplotype
networks and sequence size histogragms), and summarises the data. Two directories
for storing these outputs created automatically,"./network_diagrams" and
"./histograms". Where only representative haplotypes uploaded, rather than
individual samples, populations recorded for later inspection.

Unaligned sequence files are read in, aligned sequences are written out as
files.  Histogrmas of the orignal sequence length range and new aligned
sequence length is drawn and stored out.  Haplotype networks are also drawn
and writen out.  Summary data are recorded. If only one version of each
haplotype ever found, assumed that accessions might represent haplotypes
not sample frequency, populations flagged for further inspection.
}
