% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_genbank_df.R
\name{build_genbank_df}
\alias{build_genbank_df}
\title{Builds a reference data frame with details of all available genes associated
with each accession investigated}
\usage{
build_genbank_df(accession_file_name)
}
\arguments{
\item{accession_file_name}{user provided file name which lists accession
numbers. Should be a .csv or .txt file without a header.}
}
\description{
\code{build_genbank_df}
}
\details{
This function extracts and collates details of all available genes (as well
as the species scientific name) for each accession given.

Function designed to construct a raw reference data frame of all sequence
information available for each of the accession numbers given. Uses functions
\code{load_accession_list}, \code{get_sample_details}, and
\code{get_gene_details} to extract information from NCBI website. Output will
need downstream processing to remove duplicated pieces of information and to
clarify any nomlencature issues for example.
}
