% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_poss_synyms.R
\name{check_poss_synyms}
\alias{check_poss_synyms}
\title{Checks if scientific name for sample is likely a species, or subspecies name}
\usage{
check_poss_synyms(data)
}
\arguments{
\item{data}{data frame with `$sci_nam` column (expected to be `GB_with_SeqDat`,
the data frame result of the pipeline so far).}
}
\description{
\code{check_poss_synyms}
}
\details{
Function takes data frame with column $sci_nam and returns a list of
scientific names that are longer than 2 words. Also writes out a .csv file of
these names.

Sequence data can be uploaded to GenBank with scientific names that may be
non-standard/of different taxonomic rank, filtering based on species name will
therefore not work unless names are standardised. This function takes any
instances of names longer than 2 words, outputs the result as an object and a
.csv which can then be edited by the user.
}
