% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_low_sample_size.R
\name{drop_low_sample_size}
\alias{drop_low_sample_size}
\title{Drops populations from main data frame when sample size is less than threshold value}
\usage{
drop_low_sample_size(info_df, min_sample)
}
\arguments{
\item{info_df}{expected to be data frame of same structure as info_df built
under earlier functions. Expected column names are `spp_name`, `haplo_number`,
`sequence_length`, `sample_size`, `max_step`}

\item{min_sample}{minimum number of haplotypes before user considers data will
be uninformative}
}
\description{
\code{drop_low_sample_size}
}
\details{
Remove populations that have a sample size of less than a threshold value.

Find and drop from master list populations that have insufficient numbers of
haplotypes to be valuable for further downstream analysis - value set by user.
}
