% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_low_sequence_length.R
\name{drop_low_sequence_length}
\alias{drop_low_sequence_length}
\title{Drops populations from main data frame when sequence length is less than
threshold value}
\usage{
drop_low_sequence_length(info_df, min_length)
}
\arguments{
\item{info_df}{expected to be data frame of same structure as info_df built
under earlier functions. Expected column names are `spp_name`, `haplo_number`,
`sequence_length`, `sample_size`, `max_step`}

\item{min_length}{minimum length of sequence before user considers data will
be uninformative}
}
\description{
\code{drop_low_sequence_length}
}
\details{
Remove populations that, after alignment, only have sequence data of length
less than a threshold value.

Find and drop from master list populations that have insufficent bp to be
valuable for further downstream analsysis - value set by user.
}
