% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_details.R
\name{export_details}
\alias{export_details}
\title{Exports downloaded details into seperate .csv files for each species}
\usage{
export_details(data, file_path = getwd())
}
\arguments{
\item{data}{data frame with $sci_nam column (expected to be GB_with_SeqDat,
the data frame result of the pipeline so far).}

\item{file_path}{path to newly created .csv file, defaults to working directory}
}
\description{
\code{export_details}
}
\details{
Function takes data frame with column $sci_nam and splits according to species
name before exporting sample details as individual .csv files.

Main input data frame is split by species (according to $sci_nam column). For
each species a separate .csv file is then written out containing all the
information downloaded and stored about each accession.
}
