% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sequences.R
\name{export_sequences}
\alias{export_sequences}
\title{Exports downloaded details into seperate .csv files for each species}
\usage{
export_sequences(data)
}
\arguments{
\item{data}{data frame with `$sci_nam`, `$sequence`, and `$accession_version`
column (expected to be `GB_with_SeqDat`, the data frame result of the pipeline
so far).}
}
\description{
\code{export_sequences}
}
\details{
Function takes data frame with column `$sci_nam` and splits according to species
name before exporting sequence data from each species as seperate .fasta files.

Input data frame is split by species (according to `$sci_nam` column). For each
species a separate .fasta file is then written out containing the raw sequence
data downloaded from each accession.
}
