% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_of_interest.R
\name{gene_of_interest}
\alias{gene_of_interest}
\title{Extracts data frame rows associated with a specified gene of interest}
\usage{
gene_of_interest(gene, data)
}
\arguments{
\item{gene}{gene of interest written as character string. Must be in
"standardised" format used to simplify data in \code{standardise_gene_names}
function.}

\item{data}{data frame (expected to be `GB_data`, the data frame result of
the pipeline so far).}
}
\description{
\code{gene_of_interest}
}
\details{
Function takes data frame, likely built from all available data associated
with the original input accession numbers, and returns only data relevant to
specified gene of interest.

This function filters all data collected and curated so far from the original
accession list down to just information associated with a nominated gene of
interest.
}
