% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GB_sequence_data.R
\name{get_GB_sequence_data}
\alias{get_GB_sequence_data}
\title{Builds a curated data frame with sample details and sequence data for a
specified gene given accession numbers.}
\usage{
get_GB_sequence_data(accessions_of_interest, gene, new_names_file)
}
\arguments{
\item{accessions_of_interest}{user provided file which lists accession numbers.
Should be a .csv or .txt file without a header.}

\item{gene}{gene of interest written as character string.}

\item{new_names_file}{csv file of species names that have been flagged as
problematic and the replacement name to be used, needed for
`standadises_spp_name` function}
}
\description{
\code{get_GB_sequence_data}
}
\details{
This function extracts and collates details of the sample, as well as the
sequence data for a selected gene, from GenBank for each accession number given.

Function designed to construct a comprehensive reference data frame of sample
and sequence data available for the gene of interest from each of the
accession numbers given. Uses functions \code{load_accession_list},
\code{get_sample_details}, and \code{get_gene_details} to extract information
from NCBI website. As well as \code{standardise_gene_names} and
\code{remove_duplicates} to clean up data frame.
}
