% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_details.R
\name{get_gene_details}
\alias{get_gene_details}
\title{Gets details of all available genes associated with accession in question}
\usage{
get_gene_details(feature_name, freq, full_xmlTree)
}
\arguments{
\item{feature_name}{Names of genes/features found associated with the accession.
Function \code{get_sample_details} outputs `feature_name` value.}

\item{freq}{Frequency at which each of the `feature_name` genes/features were
found within the XML tree for this accession. Function \code{get_sample_details}
outputs a `freq` value associated with `feature_name` value}

\item{full_xmlTree}{xml tree downloaded from NCBI website and parsed to R
structure with \code{xmlInternalTreeParse} - should be provided by earlier
stages of the \code{build_genbank_df} function.}
}
\description{
\code{get_gene_details}
}
\details{
This function extracts details and sequence data of genes available for each
accession from an XML tree

Function designed to sit within \code{build_genbank_df}. In general, the
input files should be provided by earlier stages of the
\code{build_genbank_df} function. Extracts details about the accession being
investigated, such as species name.
}
