% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_details.R
\name{get_sample_details}
\alias{get_sample_details}
\title{Gets details of individual samples}
\usage{
get_sample_details(full_xmlTree)
}
\arguments{
\item{full_xmlTree}{xml tree downloaded from NCBI website and parsed to R
structure with \code{xmlInternalTreeParse} by earlier stages of the
\code{build_GB_dataframe} function.}
}
\description{
\code{get_sample_details}
}
\details{
This function gets basic background details of the sample from the xml tree
input

Function designed to sit within \code{build_GB_dataframe}. In general, the
full_xmlTree should be provided by earlier stages of the
\code{build_GB_dataframe} function. Extracts details about the accession,
such as species name.
}
