% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapfreq_from_paper.R
\name{hapfreq_from_paper}
\alias{hapfreq_from_paper}
\title{Find instances where all accessions from one paper are unique haplotypes}
\usage{
hapfreq_from_paper(freq_by_study, new_rownam, max_haps_found_together)
}
\arguments{
\item{freq_by_study}{matrix created by \code{haploFreq} function (pegas) using
categorical variable formed of first 5 characters of accession number.}

\item{new_rownam}{character string of accession versions with first five
characters repeated and pasted with an underscore in front of full string to
create a two part row name ID}

\item{max_haps_found_together}{Threshold value for number of times a haplotype
can be found repeated in the uploads from one study before paper is assumed to
have submitted ever sample rather than unique haplotypes.}
}
\description{
\code{hapfreq_from_paper}
}
\details{
This function finds the assocatied paper in instances where every accession
from one study represents a novel haplotype, likely indicating data is not at
sampled frequency.

Where every accession is a novel haplotype for the given species/gene this
function finds the associated paper.  In such instances the original paper
will need to be read to find values to multiple haplotypes by in order to
reach original sampled frequency.
}
