% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_outlier.R
\name{identify_outlier}
\alias{identify_outlier}
\title{Identifies orignial accession number of given haplotype index number}
\usage{
identify_outlier(haplotypes, fasta_data, outlier_number)
}
\arguments{
\item{haplotypes}{list of haplotypes created by pegas \code{haplotype}}

\item{fasta_data}{fasta format sequence data as read in by \code{read.fas}}

\item{outlier_number}{index number of the haplotype in network that is considered
to be an outlier, index numbers are attributed by pegas \code{haploNet}}
}
\description{
\code{identify_outlier}
}
\details{
Uses fasta format sequence data, associated list of haplotypes, and index
number of interest to extract orignal acession number of outlying sample.

Function uses fasta format sequence data, associated list of haplotypes
created by pegas \code{haplotype}, and a given index number of interest to
extract orignal acession number of outlying sample.
}
