% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_accession_list.R
\name{load_accession_list}
\alias{load_accession_list}
\title{Loads list of accession numbers}
\usage{
load_accession_list(accession_file_name)
}
\arguments{
\item{accession_file_name}{user provided file which lists accession numbers.
Should be a .csv or .txt file. No header.}
}
\description{
\code{load_accession_list}
}
\details{
This function reads in list of accession numbers provided by the user and
cleans up formatting (such as differences between "accession number" and
"accession version" (e.g. xxxx vs xxxx.1)).
Also any accession numbers starting with `[XX]_` are reviewed. These are RefSeq
accessions and are known to frequently be duplicates of other accessions in
GenBank. If the "original" accession is in the accession list already the
`[XX]_` accession number will be removed.
}
