% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_dropped.R
\name{outliers_dropped}
\alias{outliers_dropped}
\title{Identifies orignial accession number of given haplotype index number}
\usage{
outliers_dropped(max_mutations, info_df)
}
\arguments{
\item{max_mutations}{user defined value for number of mutatiosn allowed before
haplotype/sample considered an outlier}

\item{info_df}{expected to be info_df or structured as info_df built from
earlier functions. Either read in file or use object of correct format.}
}
\description{
\code{outliers_dropped}
}
\details{
Finds aligned sequence data for the populations with samples that are over a
user defined value for maximum number of mutations on a branch before being
considered to be an outlier.  Haplotype networks are plotted and outlier
sequences are dropped from fasta files.

Function identifies populations that have samples seperated by a number of
mutations greater than a user defined value for being considered an outlier.
Outlier sequences are dropped from fasta files, haplotype networks repolotted
and new files written out with prefix `new_`.
}
