% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Removes multiple copies of one gene}
\usage{
remove_duplicates(df_to_update)
}
\arguments{
\item{df_to_update}{data frame with a column name `$gene_name`. Expected to be
`GB_data`, the result of pipeline so far.}
}
\description{
\code{remove_duplicates}
}
\details{
Any instances of one gene being recorded multiple times for a single accession
number are found and reduced to a unique occourance.

Due to the structure of NCBI xml tree, after \code{standardise_gene_names} has
been used, there are likely to be multiple occourances of one gene associated
with a single accession number, these need to be removed.
}
