% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_spp_names.R
\name{standardise_spp_names}
\alias{standardise_spp_names}
\title{Updates any species names user wants to alter}
\usage{
standardise_spp_names(data, new_names_file)
}
\arguments{
\item{data}{data frame with `$sci_nam` column (expected to be `GB_with_SeqDat`,
the data frame result of the pipeline so far).}

\item{new_names_file}{A csv file name. csv file with two columns. First column
is existing scientific name, second column is name to replace existing
character string with. Structure built on exported "poss_synonyms.csv" file.}
}
\description{
\code{standardise_spp_names}
}
\details{
Function takes data frame with column $sci_nam and uses a .csv file containing
existing names and desired replacement names to standise species nomenclature
across all accessions.

Sequence data can be uploaded to GenBank with scientific names that may be
non-standard/of different taxonomic rank, filtering based on species name will
therefore not work unless names are standardised. This function reads in a .csv
file of old and user-defined replacement species names, these are then used to
update the data$sci_nam column.
}
