% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_info_df.R
\name{updating_info_df}
\alias{updating_info_df}
\title{Multiplies sequence data up to sampled frequency, aligns and summarises data}
\usage{
updating_info_df(original_df, new_df, file_name = "Info_df.csv")
}
\arguments{
\item{original_df}{expected to be info_df or structured as info_df from
earlier functions}

\item{new_df}{expected to be mag_df data frame or structure as mag_df
constructed in earlier function.}

\item{file_name}{defaults to 'Info_df.csv' but user can set name if desired,
must include file extension, e.g. '.csv'}
}
\description{
\code{updating_info_df}
}
\details{
Takes newly created data frame and merges with exisiting data frame of same
structure to ensure one master data frame is maintained.

Function reads in or takes newly created data frame and adds information to
existing data frame of same structure, the combined data frame is the written
out as a .csv file as well as being returned by the function.  New data frame
is based on files not previously processed because papers uploaded haplotypes
not sample frequency, so data needed additional processing before inclusion.
This function ensures one master data frame is maintained.
}
